Main ist jetzt entry point

+++ src/main/java/com/atdit/booking/financialdata/Credit.java
@@ -1,6 +1,6 @@
 package com.atdit.booking.financialdata;
 
-public record Credit(String description, int amount, float interestRate, int monthlyPayment, int remainingSum) {
+public record Credit(String description, int amount, double interestRate, int monthlyPayment, int remainingSum) {
 
     public int amountOwed() {
         return amount - remainingSum;
